assignment in internship
the assignment make me use the scala that I learn such as syntax. the code in Assignment 1.1 will create the data from random and covert to Csv file in 1,000,000 rows. and Assignment 1.2 will bring the part of data from Assignment 1.1 and create a new data and convert to Json file.